fix(admin-ui): align agent evidence freshness
