+++ .github/scripts/check-test-intelligence-ecosystem.py
@@ -206,9 +206,13 @@ def check(root: Path) -> list[str]:
     if "may contain sensitive browser data" not in tests_page:
         errors.append("Admin UI does not expose the diagnostic privacy warning")
     synthetic_route = text(root / "openbank-admin-ui/src/app/api/test-intelligence/route.ts")
+    freshness = text(root / "openbank-admin-ui/src/lib/test-intelligence-freshness.ts")
     for needle in ("function enforceRuntimeFreshness", "runtimeFreshnessState(item.state, item.observedAt)", "staleEvidence: evidence.filter"):
-        if needle not in synthetic_route:
+        if needle not in freshness:
             errors.append(f"running Admin UI can keep an expired successful snapshot green: {needle}")
+    agent_route = text(root / "openbank-admin-ui/src/app/api/test-intelligence/agents/route.ts")
+    if "runtimeFreshnessState(item.state as EvidenceState, item.observedAt)" not in agent_route:
+        errors.append("AI agent can analyze an expired successful snapshot as current evidence")
     for needle in ("kube_cronjob_status_last_schedule_time", "kube_cronjob_status_last_successful_time", "kube_job_status_failed"):
         if needle not in synthetic_route:
             errors.append(f"synthetic runtime projection lost its verified Kubernetes signal: {needle}")
