+++ openbank-admin-ui/src/app/api/test-intelligence/route.ts
@@ -5,6 +5,7 @@ import { promises as fs } from 'fs'
 import path from 'path'
 import { NextResponse } from 'next/server'
 import type { EvidenceState, TestIntelligenceReport } from '@/lib/types/test-intelligence'
+import { enforceRuntimeFreshness } from '@/lib/test-intelligence-freshness'
 
 export const dynamic = 'force-dynamic'
 
@@ -25,67 +26,6 @@ const emptyReport = (error: string): TestIntelligenceReport => ({
 type PrometheusVector = { status?: string; data?: { result?: { value?: [number, string] }[] } }
 type PrometheusLabelVector = { status?: string; data?: { result?: { metric?: Record<string, string>; value?: [number, string] }[] } }
 type TempoSearch = { traces?: Array<{ traceID?: string }> }
-const evidenceFreshnessMs = () => {
-  const days = Number(process.env.OPENBANK_TEST_INTELLIGENCE_STALE_AFTER_DAYS ?? 14)
-  return (Number.isFinite(days) && days > 0 ? days : 14) * 86_400_000
-}
-
-function runtimeFreshnessState(state: EvidenceState, observedAt: string | null | undefined): EvidenceState {
-  if (state === 'failed' || state === 'blocked' || state === 'unknown' || state === 'not-run') return state
-  const observed = Date.parse(observedAt ?? '')
-  if (!Number.isFinite(observed)) return 'not-run'
-  return Date.now() - observed > evidenceFreshnessMs() ? 'stale' : state
-}
-
-function enforceRuntimeFreshness(report: TestIntelligenceReport): TestIntelligenceReport {
-  const components = report.components.map(component => ({
-    ...component,
-    evidence: component.evidence.map(item => ({
-      ...item, state: runtimeFreshnessState(item.state, item.observedAt),
-    })),
-    coverage: {
-      ...component.coverage,
-      state: runtimeFreshnessState(component.coverage.state, component.coverage.observedAt),
-    },
-  }))
-  const evidence = components.flatMap(component => component.evidence)
-  return {
-    ...report,
-    components,
-    contracts: report.contracts.map(item => ({
-      ...item, state: runtimeFreshnessState(item.state, item.observedAt),
-    })),
-    mutations: report.mutations.map(item => ({
-      ...item, state: runtimeFreshnessState(item.state, item.observedAt),
-    })),
-    performance: report.performance.map(item => ({
-      ...item, state: runtimeFreshnessState(item.state, item.observedAt),
-    })),
-    syntheticJourneys: report.syntheticJourneys.map(journey => ({
-      ...journey,
-      ...(journey.ci ? { ci: {
-        ...journey.ci, state: runtimeFreshnessState(journey.ci.state, journey.ci.observedAt),
-      } } : {}),
-    })),
-    clientExperiences: report.clientExperiences.map(client => ({
-      ...client,
-      evidence: client.evidence.map(item => ({
-        ...item, state: runtimeFreshnessState(item.state, item.observedAt),
-      })),
-    })),
-    totals: {
-      ...report.totals,
-      components: components.length,
-      componentsWithExecutionEvidence: components.filter(component => component.evidence.length > 0).length,
-      moneyPathComponents: components.filter(component => component.moneyPath).length,
-      failingEvidence: evidence.filter(item => item.state === 'failed').length,
-      missingEvidence: components.filter(component => component.evidence.length === 0).length,
-      staleEvidence: evidence.filter(item => item.state === 'stale').length,
-      unknownEvidence: evidence.filter(item => item.state === 'unknown').length,
-      unresolvedEvidence: evidence.filter(item => ['unknown', 'not-run', 'blocked'].includes(item.state)).length,
-    },
-  }
-}
 
 // Journey ids arrive from the build-time bundled report, so they are file data reaching an
 // outbound Prometheus request. Kubernetes object names are already restricted to this
